Rename orchard_flavor_ext module to shielded_data_flavor
